feat(main.rs): cast return value of Sampler function to i32

+++ src/main.rs
@@ -893,7 +893,7 @@ impl Sampler {
                     self.topp,
                     &mut self.probindex,
                     coin,
-                );
+                ) as i32;
             }
         }
